nanoAgent-langchain/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nano-agent-langchain"
version = "0.1.0"
description = "nano-agent 的 LangChain 重写：CLI、工具路由与会话持久化。"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langchain>=0.3.14",
  "langchain-core>=0.3.14",
  "langchain-openai>=0.2.14",
  "pydantic>=2.7.0",
  "pyyaml>=6.0.1",
  "tenacity>=9.0.0",
  "tavily-python>=0.5.0",
]

[project.scripts]
nano-agent-langchain = "nano_agent_langchain.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
